temp_suggestions/python/requirements.txt
gui-agents
pyautogui
pillow
requests
